strava_log/requirements.txt
requests>=2.31
python-dotenv>=1.0
click>=8.1
python-dateutil>=2.8
